move iprange to lib/types/
